+++ pkg/ccl/streamingccl/streamclient/BUILD.bazel
@@ -15,10 +15,10 @@ go_library(
         "//pkg/roachpb:with-mocks",
         "//pkg/security",
         "//pkg/sql",
+        "//pkg/sql/catalog",
         "//pkg/sql/catalog/catalogkeys",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/tabledesc",
-        "//pkg/sql/randgen",
         "//pkg/sql/rowenc",
         "//pkg/sql/sem/tree",
         "//pkg/util/hlc",
