+++ pkg/cli/flags_test.go
@@ -66,6 +66,7 @@ func TestNoLinkForbidden(t *testing.T) {
 			"testing",  // defines flags
 			"go/build", // probably not something we want in the main binary
 			"github.com/cockroachdb/cockroach/pkg/security/securitytest", // contains certificates
+			"github.com/cockroachdb/cockroach/pkg/sql/randgen",           // meant for testing code only
 		},
 		[]string{
 			"github.com/cockroachdb/cockroach/pkg/testutils", // meant for testing code only
