+++ pkg/sql/colexec/colexectestutils/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "//pkg/col/coldata",
         "//pkg/col/typeconv",
         "//pkg/settings/cluster",
+        "//pkg/sql/catalog/descpb",
         "//pkg/sql/colexec/colexecargs",
         "//pkg/sql/colexecerror",
         "//pkg/sql/colexecop",
@@ -21,10 +22,12 @@ go_library(
         "//pkg/sql/execinfrapb",
         "//pkg/sql/parser",
         "//pkg/sql/randgen",
+        "//pkg/sql/rowenc",
         "//pkg/sql/rowexec",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
         "//pkg/util/duration",
+        "//pkg/util/encoding",
         "//pkg/util/envutil",
         "//pkg/util/json",
         "//pkg/util/log",
